Review modules' import
I tried to regroup all classes from the different modules into the
top-level module, but it was a waste of time!
