+++ stoiridhtools/__init__.py
@@ -1,8 +1,2 @@
 # -*- coding: utf-8 -*-
-from .config import Config
-from .sdk import SDK
-from .versionnumber import VersionNumber
-
-__all__ = ['Config', 'SDK', 'VersionNumber']
-
 __version__ = '0.1.0'
